hypre with hip GNUmake improvements (#4514)
When compiling with HIP and HYPRE, use the `rocblas` and `rocsolver`
libraries (similar to additional libraries for CUDA + HYPRE) so these
don't have to be turned off when building HYPRE. Also, check in `lib64`
for the HYPRE library, which is where it gets put when building HYPRE
with cmake.

These modifications helped with building PeleLMeX with HYPRE support on
Frontier.