Tweak CArena Defragmentation Strategy
The previous strategy has a flaw. Suppose a CArena's initial size is small
and we have n vectors each with a size of x. Now we are resizing these
vectors one by one to size x+y, where y << x. Then we would end up with n
new allocations each with a size of 2*x+y. We have doubled the memory usage
in the end, because the unused spaces can not be combined.

In the new strategy, we only attempt to combine uncombined allocations when
the combined amount is not less than the requested amount of allocation.

We also check the malloc error code now. If it fails, we will try to free
more memory and call malloc again.