Reduce use of SuperParticle (#4526)
## Summary

In preparation of #4404, this PR reduces the reliance of SuperParticle.
The issue with getSuperParticle is that the compiler is instructed to
read in all the data of the particle into registers even if most of it
is not needed. Additionally, runtime data is not accessible through
SuperParticle.

## Additional background

The previous version of packIOData with RunOnGpu had a bug when some
particles were invalid.

## Checklist

The proposed changes:
- [x] fix a bug or incorrect behavior in AMReX
- [ ] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate

---------

Co-authored-by: Andrew Myers <atmyers@lbl.gov>