ParmParse: Remove two Aborts
ParmParse now supports names containing special characters such as `/`, `(`,
`)`, and `,` except for white space and `=` (e.g., `dAx_x/dx(x,y,t,zeval) =
12`). In fact, even before this PR, such names were already supported as
long as they are not the first entry, due to the way input parsing was
implemented. With this PR, these names are now allowed as the first entry as
well.