EB anisotropic -  set_eb_data changes (#4577)
## Summary
This PR makes the necessary changes for the volumetric centroid for
anisotropic meshes. The derivation is given in the pdf in the link
below. See section 2.3 - equations 36, 40, 44, 45 and 46.

For the sake of completion, the derivation of the volume of the EB cell
and the EB face centroid is given in section 2.1 and 2.2.


[EB_anisotropic_3D.pdf](https://github.com/user-attachments/files/21354300/EB_anisotropic_3D.pdf)