Account for metric term in fluxes due to inhomogeneous Neumann boundary conditions (#4579)
This fix targets the calculation of fluxes following a linear solve
using an MLCellABecLapT linear operator with a metric term. Previously,
there was an inconsistency between fluxes within the domain and at the
domain boundaries - the presence of the metric term was only accounted
for in the calculation of fluxes within the domain but not in the
calculation of fluxes at the domain boundary. By changing the order of
operations in the MLCellABecLapT::getFluxes method, this fix solves that
issue and fluxes are calculated consistently across the domain.