Correction to EB volume centroid for anisotropic case (#4586)
## Summary
Based on the observation by @skang67, correction for the EB volume
centroid for anisotropic meshes is made.

In `cut_face_2d`, the face centroid locations -`fcx(i,j,k,0)` is
normalized w.r.t `dy` and `fcx(i,j,k,1)` is w.r.t. `dz`. This derivation
makes the volume centroid consistent with this normalization. The
updated derivation is in the pdf linked below.


[EB_anisotropic_volume_centroid.pdf](https://github.com/user-attachments/files/21498252/EB_anisotropic_volume_centroid.pdf)

The volume of the cell needs to be corrected as well. @skang67 will
submit the PR for that.

## Additional background

## Checklist

The proposed changes:
- [x] fix a bug or incorrect behavior in AMReX
- [ ] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate

Co-authored-by: Mahesh Natarajan <nataraj2@login39.chn.perlmutter.nersc.gov>