CMake: `AMReX::Flags_FASTMATH` Interface Helper Target (#4591)
## Summary

This PR _changes no defaults_. It just refactors #4545 a bit further
prior to adding the option `AMReX_FASTMATH` in a first AMReX release,
providing a better internal implementation and a user-friendly helper
target (fully optional).

Add an interface target for the fast-math flags. This makes it easier to
re-use them downstream when AMReX is pre-compiled w/o fast-math (default
in all of our current package managers), but the main code wants to
still use fast-math on the side of the application code.

## Additional background

Follow-up to #4545, see comment
https://github.com/AMReX-Codes/amrex/pull/4545#discussion_r2245683394