Improvements to AMREX_ASSERT (#4581)
## Summary

Previously the `AMREX_ASSERT_WITH_MESSAGE` and
`AMREX_ALWAYS_ASSERT_WITH_MESSAGE` macros converted the messege to a
string using `# MSG` which resulted in strange formatting and the
printing of runtime values to not work.

Additionally, `ALWAYS_ASSERT` and `ASSERT` did not work in device code
even when compiling with `AMREX_USE_ASSERTION` due to the `#if
defined(NDEBUG)` in `Assert (...)`.

Now `ALWAYS_ASSERT` and `ASSERT` will work in device code with
`AMREX_USE_ASSERTION`, but not without so the behaviour doesn't change
of the few places that use `ALWAYS_ASSERT` on the GPU (meaning
`ALWAYS_ASSERT` still does not always assert).

Test of the assert:

https://github.com/AMReX-Codes/amrex/blob/5bd9494f9229b6a7a60cbc6f2560e5a831aaa3de/Src/Particle/AMReX_StructOfArrays.H#L122

Dev:
```
0::Assertion `pos != m_rdata_names->end()' failed, file "/users/sinnalex/hipace/build/_deps/fetchedamrex-src/Src/Particle/AMReX_StructOfArrays.H", line 122,
Msg: "Soa Real name='" + name + "' was not found components" !!!
SIGABRT
See Backtrace.0 file for details
```
PR:
```
0::Assertion `pos != m_rdata_names->end()' failed, file "/users/sinnalex/amrex/Src/Particle/AMReX_StructOfArrays.H", line 122, 
Msg: Soa Real name='wrong name' was not found components !!!
SIGABRT
See Backtrace.0 file for details
```