`AMREX_ENUM`: Cache String Map
The `getEnumNameValuePairs` logic used below a lot of
`AMREX_ENUM` functions is very slow for long-enough (e.g.,
N>70) enums.

In WarpX 1D simulations, it took up close to 50% of the runtime,
as Dave Grote found.

This caches the key-value string list on first creation, because
the type is known at compile-time.

A better implementation might skip the dictionary altogether for
functions that just need one string, like `getEnumNameString`.