SumBoundary reimplementation
In the old approach, we make a temporary copy of the data, zero out the
original data, and then perform parallel add. In the new approach that
should be more efficient, we no longer need a temporary copy of the entire
data.

We have also added an optional argument that can be used to make SumBoundary
deterministic. Note that being deterministic does not mean the data are
consistent on shared nodes. For that, one still needs to call OverrideSync,
or FillBounaryAndSync.