Adding support to memory collection of Particles and MultiFabs within a grid (#4629)
This PR adds functions to return memory usage from `Particles` and
`FabArray`/`MultiFabs` within a grid/box. The main goal is to use this
information for load balancing. We aim to consider the memory allocated
to a worker as a factor in the load-balancing algorithms for AMReX, so
there is a need to capture specific grid/box memory information.

---------

Co-authored-by: Andrew Myers <atmyers@lbl.gov>
Co-authored-by: Weiqun Zhang <WeiqunZhang@lbl.gov>