EB: Fix corner case (#4687)
When the EB cuts off a tiny corner, we may encounter situations where
five faces have an aperture of one and the remaining face with an
aperture of almost one. In such cases, roundoff errors can significantly
affect the computation of the normal and the boundary centroid. This PR
addresses the issue by treating faces that are only minimally cut as
regular faces.