Robustify file renaming (#4721)
In `amrex::UtilCreateCleanDirectory`, if a directory already exists, we
preserve it by renaming it to a directory whose name includes a string
generated by `amrex::UniqueString()`. However, this string is not
guaranteed to be unique. The behavior of `std::rename` is
implementation-defined, and it may fail if the destination already
exists. This happens rarely, but it does happen.

This PR addresses the issue by removing the existing conflicting
directory before renaming. This makes the file renaming process more
robust.