VisMF::Read: Allow different number of ghost cells (#4725)
It used to be an assertion error when the number of ghost cells in the
MultiFab is different from that in the VisMF file. We now allow them to
be different. Note that if there are more ghost cells in the MultiFab,
the extra ghost cell regions are not filled.