Redistribute Particles
When redistributing particles, we need to resize a number of vectors. This
sometimes causes out-of-memory issues for GPU runs, because the vector
resize may fragment the memory arena. To address the issue, we try to
minimize the number of memory reallocation calls by reserving space before
we unpack local and remote communication buffers. Previously, resize may
cause memory reallocation in both unpacking operations.

We have also collected all the vectors that need to be resized into one
place. This will allows to explore different strategies for how to resize a
number of vectors.