fix: SpMV CPU version with out of bounds access
Currently, when computing matrix vector multiplication on the CPU, the
`amrex::SpMV` method accesses memory of the input vector wherever the
matrix has non zero column indices. This can lead to errors since
accessing out of bounds memory is UB and can effectively lead to wrong
results.

This commit proposes first a fix checking if the column value is in the
correct range.

This commit also proposes to put the OpenMP loop one loop higher to
avoid excessive thread spawning.