`PC::make_alike<Allocator>`
Change the template default of `make_alike<>()` to use the
same allocator as the creating allocator. This is a breaking
change.

In practice, this was so far used to create a pinned allocator,
e.g., for I/O operations. It is more sensible to create the
same allocator if unspecified as the creating PC and it
also makes "modern"/future AMReX, which uses PCs with
`PolymorphicArena` allocators, shorter/cleaner in user
code.