Use GitHub App to report CI status (#4796)
If this works, we will no longer nee to have a runner waiting for the
gitlab ci results. After gitlab finishes the CI, it will use the GitHub
App we have installed to post a comment.

Note that we will not be able to test it until this is merged.