Add partitionParticles version that skips reduction (#4803)
## Summary

This PR adds an overload of partitionParticles that takes num_left as an
input to skip the reduction that would compute num_left in the original
function. This can be useful when combining the reduction with other
operations in an effort to reduce the overhead from extra kernel
launches and stream synchronizations.

## Additional background

## Checklist

The proposed changes:
- [ ] fix a bug or incorrect behavior in AMReX
- [x] add new capabilities to AMReX
- [ ] changes answers in the test suite to more than roundoff level
- [ ] are likely to significantly affect the results of downstream AMReX
users
- [ ] include documentation in the code and/or rst files, if appropriate