Add amrex::Gpu::freeAsync (#4804)
This PR adds the function `amrex::Gpu::freeAsync (Arena* arena, void*
mem)` that can be used to free memory the next time the current GPU
stream is synchronized.

This is based on #4432 but with much reduced complexity from OMP.
The interface is now opt-in and always available, instead of needing to
be enabled using runtime parameters.