Add SArena: The_Arena wrapper using Gpu::freeAsync (#4835)
## Summary

As discussed
here https://github.com/AMReX-Codes/amrex/pull/4804#issuecomment-3599493751,
this PR adds a new memory arena that is a wrapper for The_Arena but
calls  freeAsync in its free function. The new arena, called SArena for
stream-ordered arena, replaces the PArena in the implementation of
The_Async_Arena.

## Additional background

I have kept the profiling of the SArena for now. However, I'm not sure
how useful it is since the CArena is already profiled.