fix: SpMatrix::prepare_comm when all_nnz == local_nnz in parallel
This commit reverts the previous fix to SpMV and targets instead the
`SpMatrix::prepare_comm` method instead. In the case where the problem
is distributed but the non-zeros are local to each process, the previous
implementation did not shift the column values leading to accessing
unallocated memory.

This fix introduces an else clause that simply shifts the values of the
columns in this scenario similar to what happens when the non-zeros are
separated into local and remote parts.

This commit allows both the new `SpMv` test and the existing `GMRES`
case to pass.