fix: `SpMatrix::prepare_comm` in case of no communication necessary (#4844)
Currently, when computing matrix vector multiplication on the CPU, the
`amrex::SpMV` method can accesses memory of the input vector wherever
the matrix has non zero column indices if the matrix is distributed but
does not need any communication to do the multiplication. This can lead
to errors since accessing out of bounds memory is UB and can effectively
lead to completely wrong results.

This contribution proposes first a fix by adding an else clause for the
case where there are only local non-zero values in the matrix.

This commit also proposes to put the OpenMP loop one loop higher to
avoid excessive thread spawning at this local matrix multiplication.