Improve documentation of OverrideSync (#4849)
This improves the documentation of `OverrideSync`, based on an exchange
between @WeiqunZhang, @aeriforme, and myself in
https://github.com/BLAST-WarpX/warpx/pull/6438:
- Add @WeiqunZhang's example in
https://github.com/BLAST-WarpX/warpx/pull/6438#discussion_r2604801219 to
the docstring of `FabArray::OverrideSync` in Src/Base/AMReX_FabArray.H.
- Copy docstring of `FabArray::OverrideSync` over to docstring of
`MultiFab::OverrideSync` in Src/Base/AMReX_MultiFab.H for consistency.
- Add @WeiqunZhang's comment in
https://github.com/BLAST-WarpX/warpx/pull/6438#discussion_r2615123791 to
the documentation in Docs/sphinx_documentation/source/Basics.rst,
highlighting the difference between the two ways of synchronizing the
data with owners overriding non-owners.

---------

Co-authored-by: Weiqun Zhang <WeiqunZhang@lbl.gov>